+++ syskit-ut/ItemQSuite.cpp
@@ -11,20 +11,19 @@ BEGIN_NAMESPACE
 class MyItem: public ItemQ::Item
 {
 public:
-    MyItem(unsigned long id);
-    unsigned long id() const;
+    MyItem(unsigned int id);
+    unsigned int id() const;
     virtual ~MyItem();
 private:
-    unsigned long id_;
+    unsigned int id_;
 };
 
-inline unsigned long
-MyItem::id() const
+inline unsigned int MyItem::id() const
 {
     return id_;
 }
 
-MyItem::MyItem(unsigned long id):
+MyItem::MyItem(unsigned int id):
 ItemQ::Item()
 {
     id_ = id;
@@ -62,13 +61,13 @@ ItemQSuite::~ItemQSuite()
 
 void ItemQSuite::testCtor00()
 {
-    ItemQ q0(12UL /*capacity*/, 0 /*growBy*/);
+    ItemQ q0(12U /*capacity*/, 0 /*growBy*/);
     bool ok = q0.isOk() &&
         (!q0.canGrow()) &&
-        (q0.capacity() == 12UL) &&
+        (q0.capacity() == 12U) &&
         (q0.growthFactor() == 0) &&
-        (q0.initialCap() == 12UL);
-    (q0.numItems() == 0UL);
+        (q0.initialCap() == 12U);
+    (q0.numItems() == 0U);
     CPPUNIT_ASSERT(ok);
 
     ItemQ::Stat stat(q0);
@@ -81,33 +80,33 @@ void ItemQSuite::testCtor00()
         (q1.capacity() == ItemQ::MAX_CAP) &&
         (q1.growthFactor() == 0) &&
         (q1.initialCap() == ItemQ::MAX_CAP);
-    (q1.numItems() == 0UL);
+    (q1.numItems() == 0U);
     CPPUNIT_ASSERT(ok);
 }
 
 
 void ItemQSuite::testCtor01()
 {
-    ItemQ q0(123UL /*capacity*/, -1 /*growBy*/);
+    ItemQ q0(123U /*capacity*/, -1 /*growBy*/);
     bool ok = q0.isOk() &&
         q0.canGrow() &&
-        (q0.capacity() == 123UL) &&
+        (q0.capacity() == 123U) &&
         (q0.growthFactor() == -1) &&
-        (q0.initialCap() == 123UL);
-    (q0.numItems() == 0UL);
+        (q0.initialCap() == 123U);
+    (q0.numItems() == 0U);
     CPPUNIT_ASSERT(ok);
 
     ItemQ::Stat stat(q0);
     ok = (stat.usagePeak() == 0) && (stat.numFails() == 0) && (stat.numGets() == 0) && (stat.numPuts() == 0);
     CPPUNIT_ASSERT(ok);
 
-    ItemQ q1(23UL /*capacity*/, 45 /*growBy*/);
+    ItemQ q1(23U /*capacity*/, 45 /*growBy*/);
     ok = q1.isOk() &&
         q1.canGrow() &&
-        (q1.capacity() == 23UL) &&
+        (q1.capacity() == 23U) &&
         (q1.growthFactor() == 45) &&
-        (q1.initialCap() == 23UL);
-    (q1.numItems() == 0UL);
+        (q1.initialCap() == 23U);
+    (q1.numItems() == 0U);
     CPPUNIT_ASSERT(ok);
 }
 
@@ -120,7 +119,7 @@ void ItemQSuite::testDtor00()
     for (;;)
     {
         ItemQ::Item* item = new ItemQ::Item();
-        if (!q.put(item, 0UL /*timeoutInMsecs*/))
+        if (!q.put(item, 0U /*timeoutInMsecs*/))
         {
             break;
         }
@@ -135,7 +134,7 @@ void ItemQSuite::testDtor00()
 
 //
 // Interfaces under test:
-// - bool ItemQ::expedite(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::expedite(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testExpedite00()
 {
@@ -145,7 +144,7 @@ void ItemQSuite::testExpedite00()
     q.resetStat();
     ItemQ::Item* item;
     bool ok = true;
-    for (unsigned long i = q.capacity(); i > 0; --i)
+    for (unsigned int i = q.capacity(); i > 0; --i)
     {
         item = new ItemQ::Item();
         if (!q.expedite(item, ItemQ::ETERNITY))
@@ -158,13 +157,13 @@ void ItemQSuite::testExpedite00()
 
     // Try enqueuing when queue is full.
     item = new ItemQ::Item();
-    ok = (!q.expedite(item, 0UL /*timeoutInMsecs*/));
+    ok = (!q.expedite(item, 0U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
     item = new ItemQ::Item();
-    ok = (!q.expedite(item, 12UL /*timeoutInMsecs*/));
+    ok = (!q.expedite(item, 12U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
 
-    while (q.get(item, 0UL /*timeoutInMsecs*/))
+    while (q.get(item, 0U /*timeoutInMsecs*/))
     {
         ItemQ::Item::release(item);
     }
@@ -178,7 +177,7 @@ void ItemQSuite::testExpedite00()
 
 //
 // Interfaces under test:
-// - bool ItemQ::expedite(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::expedite(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testExpedite01()
 {
@@ -187,13 +186,13 @@ void ItemQSuite::testExpedite01()
     // Try enqueuing when queue is not full. Don't wait forever.
     q.resetStat();
     ItemQ::Item* item = new ItemQ::Item();
-    bool ok = q.expedite(item, 0UL /*timeoutInMsecs*/);
+    bool ok = q.expedite(item, 0U /*timeoutInMsecs*/);
     CPPUNIT_ASSERT(ok);
     item = new ItemQ::Item();
-    ok = q.expedite(item, 12UL /*timeoutInMsecs*/);
+    ok = q.expedite(item, 12U /*timeoutInMsecs*/);
     CPPUNIT_ASSERT(ok);
 
-    while (q.get(item, 0UL /*timeoutInMsecs*/))
+    while (q.get(item, 0U /*timeoutInMsecs*/))
     {
         ItemQ::Item::release(item);
     }
@@ -206,7 +205,7 @@ void ItemQSuite::testExpedite01()
 
 //
 // Interfaces under test:
-// - bool ItemQ::get(Item*& item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::get(Item*& item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testGet00()
 {
@@ -214,9 +213,9 @@ void ItemQSuite::testGet00()
 
     // Try dequeuing when queue is empty.
     ItemQ::Item* item;
-    bool ok = (!q.get(item, 0UL /*timeoutInMsecs*/));
+    bool ok = (!q.get(item, 0U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
-    ok = (!q.get(item, 12UL /*timeoutInMsecs*/));
+    ok = (!q.get(item, 12U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
 
     ItemQ::Stat stat(q);
@@ -228,17 +227,17 @@ void ItemQSuite::testGet00()
 //
 // Interfaces under test:
 // - bool ItemQ::apply(cb0_t cb, void* arg=0) const;
-// - bool ItemQ::get(Item*& item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::get(Item*& item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testGet01()
 {
     ItemQ q(32 /*capacity*/, 0 /*growBy*/);
 
-    unsigned long id;
+    unsigned int id;
     for (id = 0;; ++id)
     {
         MyItem* item = new MyItem(id);
-        if (!q.put(item, 0UL /*timeoutInMsecs*/))
+        if (!q.put(item, 0U /*timeoutInMsecs*/))
         {
             break;
         }
@@ -252,10 +251,10 @@ void ItemQSuite::testGet01()
 
     // Make sure items are queued FIFO via put().
     ok = true;
-    for (unsigned long i = 0; i < id; ++i)
+    for (unsigned int i = 0; i < id; ++i)
     {
         ItemQ::Item* item;
-        if ((!q.get(item, 0UL /*timeoutInMsecs*/)) || ((dynamic_cast<const MyItem*>(item))->id() != i))
+        if ((!q.get(item, 0U /*timeoutInMsecs*/)) || ((dynamic_cast<const MyItem*>(item))->id() != i))
         {
             ok = false;
             break;
@@ -273,18 +272,18 @@ void ItemQSuite::testGet01()
 
 //
 // Interfaces under test:
-// - bool ItemQ::expedite(Item* item, unsigned long timeoutInMsecs=ETERNITY);
-// - bool ItemQ::get(Item*& item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::expedite(Item* item, unsigned int timeoutInMsecs=ETERNITY);
+// - bool ItemQ::get(Item*& item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testGet02()
 {
     ItemQ q(32 /*capacity*/, 0 /*growBy*/);
 
-    unsigned long id;
+    unsigned int id;
     for (id = 0;; ++id)
     {
         MyItem* item = new MyItem(id);
-        if (!q.expedite(item, 0UL /*timeoutInMsecs*/))
+        if (!q.expedite(item, 0U /*timeoutInMsecs*/))
         {
             break;
         }
@@ -292,11 +291,11 @@ void ItemQSuite::testGet02()
 
     // Make sure messages are queued LIFO via expedite().
     bool ok = true;
-    for (long i = id - 1; i >= 0; --i)
+    for (int i = id - 1; i >= 0; --i)
     {
         ItemQ::Item* item;
-        if ((!q.get(item, 0UL /*timeoutInMsecs*/)) ||
-            ((dynamic_cast<const MyItem*>(item))->id() != static_cast<unsigned long>(i)))
+        if ((!q.get(item, 0U /*timeoutInMsecs*/)) ||
+            ((dynamic_cast<const MyItem*>(item))->id() != static_cast<unsigned int>(i)))
         {
             ok = false;
             break;
@@ -314,14 +313,14 @@ void ItemQSuite::testGet02()
 
 //
 // Interfaces under test:
-// - bool ItemQ::put(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::put(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testGrow00()
 {
     ItemQ q(16 /*capacity*/, -1 /*growBy*/);
 
     bool ok = true;
-    for (unsigned long i = q.capacity(); i > 0; --i)
+    for (unsigned int i = q.capacity(); i > 0; --i)
     {
         if (!q.put(0 /*item*/, ItemQ::ETERNITY))
         {
@@ -333,12 +332,12 @@ void ItemQSuite::testGrow00()
 
     size_t oldCap = q.capacity();
     size_t newCap = oldCap * 2;
-    ok = q.put(0 /*item*/, 0UL /*timeoutInMsecs*/) && (q.capacity() == newCap) && (q.initialCap() == oldCap);
+    ok = q.put(0 /*item*/, 0U /*timeoutInMsecs*/) && (q.capacity() == newCap) && (q.initialCap() == oldCap);
     CPPUNIT_ASSERT(ok);
 
     ItemQ::Item* item;
     size_t count = q.numItems();
-    while (q.get(item, 0UL /*timeoutInMsecs*/));
+    while (q.get(item, 0U /*timeoutInMsecs*/));
 
     ItemQ::Stat stat(q);
     ok = (stat.usagePeak() == count) && (stat.numFails() == 0) && (stat.numGets() == count) && (stat.numPuts() == count);
@@ -348,14 +347,14 @@ void ItemQSuite::testGrow00()
 
 //
 // Interfaces under test:
-// - bool ItemQ::expedite(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::expedite(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testGrow01()
 {
     ItemQ q(24 /*capacity*/, 12 /*growBy*/);
 
     bool ok = true;
-    for (unsigned long i = q.capacity(); i > 0; --i)
+    for (unsigned int i = q.capacity(); i > 0; --i)
     {
         if (!q.expedite(0 /*item*/, ItemQ::ETERNITY))
         {
@@ -367,12 +366,12 @@ void ItemQSuite::testGrow01()
 
     size_t oldCap = q.capacity();
     size_t newCap = oldCap + 12;
-    ok = q.expedite(0 /*item*/, 0UL /*timeoutInMsecs*/) && (q.capacity() == newCap) && (q.initialCap() == oldCap);
+    ok = q.expedite(0 /*item*/, 0U /*timeoutInMsecs*/) && (q.capacity() == newCap) && (q.initialCap() == oldCap);
     CPPUNIT_ASSERT(ok);
 
     ItemQ::Item* item;
     size_t count = q.numItems();
-    while (q.get(item, 0UL /*timeoutInMsecs*/));
+    while (q.get(item, 0U /*timeoutInMsecs*/));
 
     ItemQ::Stat stat(q);
     ok = (stat.usagePeak() == count) && (stat.numFails() == 0) && (stat.numGets() == count) && (stat.numPuts() == count);
@@ -386,7 +385,7 @@ void ItemQSuite::testGrow01()
 void ItemQSuite::testGrow02()
 {
     ItemQ q(12345 /*capacity*/, 9000 /*growBy*/);
-    while (q.put(0 /*item*/, 0UL /*timeoutInMsecs*/));
+    while (q.put(0 /*item*/, 0U /*timeoutInMsecs*/));
 
     ItemQ::Stat stat(q);
     bool ok = (q.capacity() == ItemQ::MAX_CAP) &&
@@ -401,7 +400,7 @@ void ItemQSuite::testGrow02()
 
 //
 // Interfaces under test:
-// - bool ItemQ::put(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::put(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testPut00()
 {
@@ -410,7 +409,7 @@ void ItemQSuite::testPut00()
     // Try enqueuing when queue is not full. Wait forever.
     ItemQ::Item* item;
     bool ok = true;
-    for (unsigned long i = q.capacity(); i > 0; --i)
+    for (unsigned int i = q.capacity(); i > 0; --i)
     {
         item = new ItemQ::Item();
         if (!q.put(item, ItemQ::ETERNITY))
@@ -423,13 +422,13 @@ void ItemQSuite::testPut00()
 
     // Try enqueuing when queue is full.
     item = new ItemQ::Item();
-    ok = (!q.put(item, 0UL /*timeoutInMsecs*/));
+    ok = (!q.put(item, 0U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
     item = new ItemQ::Item();
-    ok = (!q.put(item, 12UL /*timeoutInMsecs*/));
+    ok = (!q.put(item, 12U /*timeoutInMsecs*/));
     CPPUNIT_ASSERT(ok);
 
-    while (q.get(item, 0UL /*timeoutInMsecs*/))
+    while (q.get(item, 0U /*timeoutInMsecs*/))
     {
         ItemQ::Item::release(item);
     }
@@ -443,7 +442,7 @@ void ItemQSuite::testPut00()
 
 //
 // Interfaces under test:
-// - bool ItemQ::put(Item* item, unsigned long timeoutInMsecs=ETERNITY);
+// - bool ItemQ::put(Item* item, unsigned int timeoutInMsecs=ETERNITY);
 //
 void ItemQSuite::testPut01()
 {
@@ -451,13 +450,13 @@ void ItemQSuite::testPut01()
 
     // Try enqueuing when queue is not full. Don't wait forever.
     ItemQ::Item* item = new ItemQ::Item();
-    bool ok = q.put(item, 0UL /*timeoutInMsecs*/);
+    bool ok = q.put(item, 0U /*timeoutInMsecs*/);
     CPPUNIT_ASSERT(ok);
     item = new ItemQ::Item();
-    ok = q.put(item, 12UL /*timeoutInMsecs*/);
+    ok = q.put(item, 12U /*timeoutInMsecs*/);
     CPPUNIT_ASSERT(ok);
 
-    while (q.get(item, 0UL /*timeoutInMsecs*/))
+    while (q.get(item, 0U /*timeoutInMsecs*/))
     {
         ItemQ::Item::release(item);
     }
